feat: Add CozyUrl to contact row
