+++ src/components/ContactsList/Contacts/ContactCozy.jsx
@@ -1,0 +1,13 @@
+import React from 'react'
+import PropTypes from 'prop-types'
+const ContactCozy = ({ cozyUrl }) => (
+  <div className="contact-cozyurl">{cozyUrl}</div>
+)
+ContactCozy.propTypes = {
+  cozyUrl: PropTypes.string
+}
+ContactCozy.defaultProps = {
+  cozyUrl: '—'
+}
+
+export default ContactCozy
