+++ src/components/ContactsList/Contacts/ContactEmail.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+import PropTypes from 'prop-types'
+const ContactEmail = ({ email }) => <div className="contact-email">{email}</div>
+ContactEmail.propTypes = {
+  email: PropTypes.string
+}
+ContactEmail.defaultProps = {
+  email: '—'
+}
+
+export default ContactEmail
