+++ src/styles/contacts.styl
@@ -78,6 +78,10 @@ $reset-list
     @extend $table-cell
     flex 0 0 12rem
 
+.contact-cozyurl
+    @extend $table-cell
+    flex 0 0 12rem
+
 .contact-email
     @extend $table-cell
     flex-basis 30%
@@ -85,6 +89,7 @@ $reset-list
 +small-screen()
     .contact-identity
     .contact-phone
+    .contact-cozyurl
     .contact-email
         flex-basis auto
         padding .3rem 1rem
